Remove most of grunt, refactor to use npm scripts
