+++ .vscode/settings.json
@@ -26,5 +26,22 @@
   "typescript.tsdk": "./node_modules/typescript/lib",
   "editor.codeActionsOnSave": {
     "source.fixAll.eslint": true
-  }
+  },
+  "workbench.colorCustomizations": {
+    "activityBar.background": "#1d87d7",
+    "activityBar.activeBackground": "#1d87d7",
+    "activityBar.activeBorder": "#f4abd4",
+    "activityBar.foreground": "#e7e7e7",
+    "activityBar.inactiveForeground": "#e7e7e799",
+    "activityBarBadge.background": "#f4abd4",
+    "activityBarBadge.foreground": "#15202b",
+    "titleBar.activeBackground": "#176baa",
+    "titleBar.inactiveBackground": "#176baa99",
+    "titleBar.activeForeground": "#e7e7e7",
+    "titleBar.inactiveForeground": "#e7e7e799",
+    "statusBar.background": "#176baa",
+    "statusBarItem.hoverBackground": "#1d87d7",
+    "statusBar.foreground": "#e7e7e7"
+  },
+  "peacock.color": "#176BAA"
 }
