--- Excalibur.sublime-project
@@ -1,55 +1,0 @@
-{
-	"folders":
-	[
-		{
-			"path": ".",
-			"file_exclude_patterns" : [
-				"src/spec/*.js",
-				"sandbox/*.js",
-            "src/engine/*.js",
-            "*.csproj"
-			],
-			"folder_exclude_patterns" : [
-				"js",
-            "node_modules",
-            "bin"
-			]
-		}
-
-	],
-	"settings":
-	{
-		"tab_size":3,
-		"translate_tabs_to_spaces":true,
-		"detect_indentation":false
-	},
-	"build_systems":
-	[
-		{
-			"name": "Build",
-			"cmd": ["build"],
-			"working_dir":"$project_path",
-
-			"windows":{
-				"cmd":["grunt.cmd"]
-			},
-			"linux":{
-				"cmd":["grunt"]
-			}
-		},
-		{
-			"name": "Clean",
-			"cmd": ["clean"],
-			"working_dir":"$project_path",
-
-			"windows":{
-				"cmd":["clean.bat"]
-			},
-			"linux":{
-				"cmd":["make clean"]
-			}
-		}
-
-
-	]
-}
