+++ package.json
@@ -23,19 +23,29 @@
     "simple"
   ],
   "scripts": {
-    "all": "grunt",
-    "core": "grunt core",
-    "lint": "grunt lint",
+    "all": "npm run lint && npm run core && npm run test && npm run legacy-visual",
+    "clean": "rimraf ./build/dist ./src/spec/*.js ./src/spec/*.map",
+    "travis": "npm run clean && git clone https://github.com/excaliburjs/excalibur-dist build && npm run core && npm run core:building-prod && npm run coveralls && grunt dists",
+    "start": "npm run core:watch",
+    "core": "npm run core:tsc && npm run core:copy && npm run core:bundle",
+    "core:tsc": "tsc --project src/engine/tsconfig.json --incremental true --tsBuildInfoFile .tsbuildinfo",
+    "core:copy": "copyfiles -u 2 ./src/engine/**/*.png ./src/engine/**/*.css ./src/engine/**/*.glsl ./build/dist/",
+    "core:bundle": "webpack --config webpack.config.js --devtool none --progress",
+    "core:bundle-prod": "webpack --config webpack.config.prod.js",
+    "core:watch": "npm run core:bundle -- --watch",
+    "lint": "npm run eslint && npm run eslint:spec",
     "eslint": "eslint \"src/engine/**/*.ts\"",
     "eslint:sandbox": "eslint \"sandbox/**/*.ts\"",
     "eslint:spec": "eslint \"src/spec/**/*.ts\"",
-    "travis": "grunt travis",
     "visual": "start-storybook",
-    "sandbox": "grunt sandbox",
-    "legacy-visual": "grunt visual",
-    "start": "grunt watch",
+    "sandbox:copy": "copyfiles -u 2 ./build/dist/**/*.* ./sandbox/lib/",
+    "sandbox:build": "tsc --project ./sandbox",
+    "nuget": ".\\src\\tools\\NuGet.exe pack Excalibur.nuspec -OutputDirectory .\\build\\dist -version",
+    "sandbox": "serve ./sandbox -l 3001 -n",
+    "legacy-visual": "npm run sandbox:copy && npm run sandbox:build",
     "test": "karma start",
-    "apidocs": "grunt apidocs",
+    "coveralls": "coveralls < coverage/lcov.info",
+    "apidocs": "git submodule update && node apidocs.js",
     "precommit": "pretty-quick --staged && npm run eslint && npm run eslint:spec",
     "pretty": "prettier --write \"{src,sandbox/src,sandbox/tests}/**/*.{ts,js,json,css,md}\" --config prettier.config.js"
   },
@@ -61,6 +71,8 @@
     "@typescript-eslint/parser": "2.25.0",
     "babel-loader": "8.1.0",
     "copy-webpack-plugin": "5.1.1",
+    "copyfiles": "2.2.0",
+    "coveralls": "3.0.11",
     "css-loader": "3.4.2",
     "eslint": "6.8.0",
     "eslint-plugin-jsdoc": "22.1.0",
@@ -69,15 +81,6 @@
     "grunt-build-control": "0.7.1",
     "grunt-bumpup": "0.6.3",
     "grunt-cli": "1.3.2",
-    "grunt-contrib-clean": "2.0.0",
-    "grunt-contrib-connect": "2.1.0",
-    "grunt-contrib-copy": "1.0.0",
-    "grunt-coveralls": "2.0.0",
-    "grunt-karma": "3.0.2",
-    "grunt-shell": "3.0.1",
-    "grunt-string-replace": "1.3.1",
-    "grunt-ts": "6.0.0-beta.22",
-    "grunt-webpack": "3.1.3",
     "husky": "4.2.3",
     "istanbul": "0.4.5",
     "istanbul-instrumenter-loader": "3.0.1",
@@ -94,17 +97,19 @@
     "pretty-quick": "2.0.1",
     "puppeteer": "2.1.1",
     "rimraf": "3.0.1",
+    "serve": "11.3.0",
     "source-map-support": "0.5.16",
     "sync-request": "6.1.0",
     "to-string-loader": "1.1.6",
     "travis-ci": "2.2.0",
-    "ts-loader": "6.2.1",
+    "ts-loader": "6.2.2",
     "tslint": "6.1.0",
     "typedoc": "0.17.3",
     "typescript": "3.8.3",
     "url-loader": "4.0.0",
     "webpack": "4.42.1",
-    "webpack-cli": "3.3.11"
+    "webpack-cli": "3.3.11",
+    "webpack-merge": "4.2.2"
   },
   "dependencies": {
     "core-js": "3.6.4"
