+++ sandbox/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "compilerOptions": {
+    "sourceMap": true,
+    "experimentalDecorators": true,
+    "target": "es5",
+    "module": "es2015",
+    "removeComments": false,
+    "skipLibCheck": true,
+    "lib": ["dom", "es5", "es2015.promise", "es2015.collection", "es2015.iterable"]
+  }
+}
