+++ src/engine/tsconfig.json
@@ -7,6 +7,9 @@
     "target": "es5",
     "module": "es2015",
     "declaration": true,
+    "removeComments": false,
+    "incremental": true,
+    "tsBuildInfoFile": "../../.tsbuildinfo",
     "noImplicitAny": true,
     "noFallthroughCasesInSwitch": true,
     "noImplicitReturns": true,
