+++ webpack.config.prod.js
@@ -1,0 +1,10 @@
+const merge = require('webpack-merge');
+
+const base = require('./webpack.config.js');
+
+module.exports = merge(base, {
+  mode: 'production',
+  output: {
+    filename: 'excalibur.min.js'
+  }
+});
